+++ bionc/bionc_numpy/__init__.py
@@ -12,3 +12,4 @@ from .natural_axis import Axis
 from .natural_marker import SegmentMarker, Marker
 from .natural_segment import NaturalSegment
 from .inertia_parameters import InertiaParameters
+from .interpolation_matrix import interpolate_natural_vector, to_natural_vector
