+++ tests/test_natural_segment.py
@@ -1,8 +1,8 @@
 from bionc import (
     NaturalSegment,
     SegmentMarker,
-    bionc_numpy as bionc_np,
 )
+from bionc.bionc_numpy import SegmentNaturalCoordinates
 import numpy as np
 import pytest
 
@@ -71,7 +71,7 @@ def test_marker_features():
     my_segment.add_marker(marker1)
     my_segment.add_marker(marker2)
 
-    Qi = bionc_np.SegmentNaturalCoordinates.from_components(
+    Qi = SegmentNaturalCoordinates.from_components(
         u=[1, 2, 3],
         rp=[1, 1, 3],
         rd=[1, 2, 4],
